+++ ApplicationLibCode/ProjectDataModel/Intersections/RimIntersectionCollection.h
@@ -73,7 +73,7 @@ public:
     // Visualization interface
 
     void applySingleColorEffect();
-    void updateCellResultColor( bool hasGeneralCellResult, size_t timeStepIndex );
+    void updateCellResultColor( bool hasGeneralCellResult, int timeStepIndex );
     void appendPartsToModel( Rim3dView& view, cvf::ModelBasicList* model, cvf::Transform* scaleTransform );
     void rebuildGeometry();
